// src/lib/enums/stages-of-growth.ts
export enum StagesOfGrowth {
    GERMINATION = "GERMINATION",
    TRANSPLANTING = "TRANSPLANTING",
    VEGETATIVE_GROWTH = "VEGETATIVE_GROWTH",
    HEAD_FORMATION = "HEAD_FORMATION",
    BULBING = "BULBING",
    BOLTING = "BOLTING",
    SPROUTING = "SPROUTING",
    EARLY_SPROUTING = "EARLY_SPROUTING",
    HARVEST = "HARVEST",
    MATURITY = "MATURITY",
    FLOWERING = "FLOWERING",
    FRUIT_FORMATION = "FRUIT_FORMATION",
}